+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 3.4.0 (xxxx-xx-xx)
+## 3.4.0 (2019-06-20)
 
 * Deprecate "excluded_404s" option
 * Flush loggers on `kernel.reset`